move scroll centering helpers to viewport.ts
