+++ packages/excalidraw/viewport.ts
@@ -1,7 +1,20 @@
-import { easeOut, MAX_ZOOM, MIN_ZOOM, ZOOM_STEP } from "@excalidraw/common";
-import { clamp, roundToStep } from "@excalidraw/math";
+import {
+  arrayToMap,
+  easeOut,
+  MAX_ZOOM,
+  MIN_ZOOM,
+  sceneCoordsToViewportCoords,
+  viewportCoordsToSceneCoords,
+  ZOOM_STEP,
+} from "@excalidraw/common";
+import { clamp, pointDistance, pointFrom, roundToStep } from "@excalidraw/math";
 
-import { CaptureUpdateAction } from "@excalidraw/element";
+import {
+  CaptureUpdateAction,
+  getCommonBounds,
+  getElementBounds,
+  getVisibleElements,
+} from "@excalidraw/element";
 
 import type { SceneBounds } from "@excalidraw/element";
 
@@ -11,9 +24,13 @@ import type { ExcalidrawElement } from "@excalidraw/element/types";
 import { AnimationController } from "./renderer/animation";
 import { getNormalizedZoom } from "./scene";
 
-import { centerScrollOn } from "./scene/scroll";
-
-import type { AppState, NormalizedZoomValue, Offsets } from "./types";
+import type {
+  AppState,
+  NormalizedZoomValue,
+  Offsets,
+  PointerCoords,
+  Zoom,
+} from "./types";
 
 export const SCROLL_TO_CONTENT_ANIMATION_KEY = "animateScrollToContent";
 
@@ -451,4 +468,110 @@ const animateToViewport = (
       return null;
     },
   );
+};
+
+const isOutsideViewPort = (appState: AppState, cords: Array<number>) => {
+  const [x1, y1, x2, y2] = cords;
+  const { x: viewportX1, y: viewportY1 } = sceneCoordsToViewportCoords(
+    { sceneX: x1, sceneY: y1 },
+    appState,
+  );
+  const { x: viewportX2, y: viewportY2 } = sceneCoordsToViewportCoords(
+    { sceneX: x2, sceneY: y2 },
+    appState,
+  );
+  return (
+    viewportX2 - viewportX1 > appState.width ||
+    viewportY2 - viewportY1 > appState.height
+  );
+};
+
+export const getClosestElementBounds = (
+  elements: readonly ExcalidrawElement[],
+  from: { x: number; y: number },
+): Bounds => {
+  if (!elements.length) {
+    return [0, 0, 0, 0];
+  }
+
+  let minDistance = Infinity;
+  let closestElement = elements[0];
+  const elementsMap = arrayToMap(elements);
+  elements.forEach((element) => {
+    const [x1, y1, x2, y2] = getElementBounds(element, elementsMap);
+    const distance = pointDistance(
+      pointFrom((x1 + x2) / 2, (y1 + y2) / 2),
+      pointFrom(from.x, from.y),
+    );
+
+    if (distance < minDistance) {
+      minDistance = distance;
+      closestElement = element;
+    }
+  });
+
+  return getElementBounds(closestElement, elementsMap);
+};
+
+export const centerScrollOn = ({
+  scenePoint,
+  viewportDimensions,
+  zoom,
+  offsets,
+}: {
+  scenePoint: PointerCoords;
+  viewportDimensions: { height: number; width: number };
+  zoom: Zoom;
+  offsets?: Offsets;
+}) => {
+  let scrollX =
+    (viewportDimensions.width - (offsets?.right ?? 0)) / 2 / zoom.value -
+    scenePoint.x;
+
+  scrollX += (offsets?.left ?? 0) / 2 / zoom.value;
+
+  let scrollY =
+    (viewportDimensions.height - (offsets?.bottom ?? 0)) / 2 / zoom.value -
+    scenePoint.y;
+
+  scrollY += (offsets?.top ?? 0) / 2 / zoom.value;
+
+  return {
+    scrollX,
+    scrollY,
+  };
+};
+
+export const calculateScrollCenter = (
+  elements: readonly ExcalidrawElement[],
+  appState: AppState,
+): { scrollX: number; scrollY: number } => {
+  elements = getVisibleElements(elements);
+
+  if (!elements.length) {
+    return {
+      scrollX: 0,
+      scrollY: 0,
+    };
+  }
+  let [x1, y1, x2, y2] = getCommonBounds(elements);
+
+  if (isOutsideViewPort(appState, [x1, y1, x2, y2])) {
+    [x1, y1, x2, y2] = getClosestElementBounds(
+      elements,
+      viewportCoordsToSceneCoords(
+        { clientX: appState.scrollX, clientY: appState.scrollY },
+        appState,
+      ),
+    );
+  }
+
+  const centerX = (x1 + x2) / 2;
+  const centerY = (y1 + y2) / 2;
+
+  return centerScrollOn({
+    scenePoint: { x: centerX, y: centerY },
+    viewportDimensions: { width: appState.width, height: appState.height },
+    zoom: appState.zoom,
+  });
 };
